cleanup main widget
