--- lib/main.dart
@@ -1,104 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:sketchnotes_flutter/widgets/paper_widget.dart';
-
-void main() => runApp(new MyApp());
-
-class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return new MaterialApp(
-      title: 'Sketch Notes',
-      debugShowCheckedModeBanner: false,
-      theme: new ThemeData(
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or press Run > Flutter Hot Reload in IntelliJ). Notice that the
-        // counter didn't reset back to zero; the application is not restarted.
-        primarySwatch: Colors.blue,
-      ),
-      home: new MyHomePage(title: 'Sketch Notes'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
-
-  final String title;
-
-  @override
-  _MyHomePageState createState() => new _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-
-  final _paper = PaperWidget();
-
-  @override
-  Widget build(BuildContext context) {
-    return new Scaffold(
-      appBar: new AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: new Text(widget.title),
-        actions: <Widget>[
-          IconButton(
-            icon: Icon(choices[0].icon),
-            onPressed: () {
-              _select(context, choices[0]);
-            },
-          ),
-        ],
-      ),
-      body: Column(
-        children: <Widget>[
-          Expanded(
-            child: GridPaper(
-              color: Colors.blue,
-              divisions: 1,
-              subdivisions: 1,
-              interval: 20,
-              child: _paper,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  void _select(BuildContext context, Choice choice) {
-    // Causes the app to rebuild with the new _selectedChoice.
-    setState(() {
-      // clear paper
-      print("CLEAR");
-      _paper.clear();
-    });
-  }
-
-  void _addPage() {}
-}
-
-class Choice {
-  const Choice({this.title, this.icon});
-
-  final String title;
-  final IconData icon;
-}
-
-const List<Choice> choices = const <Choice>[
-  const Choice(title: 'New', icon: Icons.clear),
-];
